requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0.post0
pytz==2024.1
jinja2==3.1.4
requests==2.32.3
urllib3==1.26.18      # ← 추가/고정 (v2 대신 v1)
certifi==2024.7.4     # 최신 루트 인증서